requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0

